🚚 Move: `search_user` module to `param` module [#63]
