+++ src/client/param/mod.rs
@@ -1,3 +1,4 @@
 //! Features for the parameters of the [`Client`](crate::client::Client) struct methods.
 
 pub mod news_stream;
+pub mod search_user;
